membuat login page, database migration

+++ database/migrations/2026_07_14_082519_add_is_read_to_messages_table.php
@@ -8,21 +8,21 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('messages', function (Blueprint $table) {
-
-            $table->boolean('is_read')
-                  ->default(false)
-                  ->after('message');
-
-        });
+        if (!Schema::hasColumn('messages', 'is_read')) {
+            Schema::table('messages', function (Blueprint $table) {
+                $table->boolean('is_read')
+                      ->default(false)
+                      ->after('message');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('messages', function (Blueprint $table) {
-
-            $table->dropColumn('is_read');
-
-        });
+        if (Schema::hasColumn('messages', 'is_read')) {
+            Schema::table('messages', function (Blueprint $table) {
+                $table->dropColumn('is_read');
+            });
+        }
     }
 };
